+++ tests/test_models.php
@@ -367,3 +367,29 @@ class InvalidRelationship2 extends Model
         ],
     ];
 }
+
+class TransactionModel extends Model
+{
+    protected static $properties = [
+        'name' => [
+            'required' => true,
+            'validate' => 'string:5',
+        ],
+    ];
+
+    protected function initialize()
+    {
+        parent::initialize();
+
+        self::deleting(function (\Pulsar\ModelEvent $modelEvent) {
+            if ('delete fail' == $modelEvent->getModel()->name) {
+                $modelEvent->stopPropagation();
+            }
+        });
+    }
+
+    protected function usesTransactions(): bool
+    {
+        return true;
+    }
+}