+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 ### Added
-- Added a DBAL driver
+- Added a DBAL driver.
+- Model operations will now be wrapped in database transactions by overriding the `Model::usesTransactions()` method. 
 
 ## 0.9.1 - 2019-11-16
 ## Fixed